Split point tests from test_geometry.py
// edited by skirpichev
